libosdp: Add doxygen comments for the file API
Also size_t and ssize_t as defined by the linux alternatives for these
file ops methods seemed a bit excessive here since we cannot be sending
that many bytes anyway.

Switch to using regular int and make sure that the parameters are
validated before calling the user provided ops methods to achieve the
same effect.

Related-to: #24

+++ src/osdp_file.c
@@ -151,7 +151,7 @@ int osdp_file_register_ops(osdp_t *ctx, int pd, struct osdp_file_ops *ops)
 }
 
 OSDP_EXPORT
-int osdp_file_tx_status(osdp_t *ctx, int pd, size_t *size, size_t *offset)
+int osdp_file_tx_status(osdp_t *ctx, int pd, int *size, int *offset)
 {
 	input_check(ctx, pd);
 	struct osdp_file *f = TO_FILE(TO_PD(ctx, pd));
